pyproject.toml
[project]
name = "applicant-resume-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fasttext-wheel>=0.9.2",
    "flask>=3.1.2",
    "matplotlib>=3.10.8",
    "mysql-connector-python>=9.5.0",
    "numpy==1.26.4",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "seaborn>=0.13.2",
    "spacy>=3.8.11",
]
